// happytrip_frontend/src/app/models/addNewBusResponseModel.ts
export interface addNewBusResponse {
    busId: number
    busName: string
    source: string
    destination: string
    boardingPoint: string
    droppingPoint: string
    departureDate: string
    departureTime: string
    amenities: string
    coachType: string
    totalAvailableSeats: string
    classAvailable: string
    bookingPolicies: string
    classWithFare: string
  }
  